[QTI-64] Add enos_environment resource
As this utility is intended to be used against resources in the cloud
via an SSH transport, often those resources will need to be configured
to allow access from the host machine running enos. While we could
always require the user to determine their public IP address and pass it
in as a variable, instead we chose to create a datasource that does that
for the user. The first output of the environment resource is the
  `public_ip_address` of the machine running provider. As more needs
  emerge we can add additional information to this datasource.

The first pass has three different external IP address resolvers. The
preferred is the myip.opendns.com resolver. If one uses their resolver
to resolve that A record it will return the IP address of the caller.
Next, we do a similar action against Google's nameservers. It acts in
much the same way except that it returns a TXT record of the IP.
Finally, if both of those fail we make an HTTPS request to AWS's checkip
endpoint.

* Make resource scaffolding creator support datasources and resources
* Create the enos_environment scaffolding
* Create the public IP address resolvers and wire them into the resource
* Add a README to the core example
* Update the core example to use the public IP address from the resource
  to create a security group, rather than requiring a user to pass in a
  security group name.
* Fix a bug in the transport utility where we were uncessarily
  recreating response structs.

Signed-off-by: Ryan Cragun <me@ryan.ec>