[QTI-96] Add --replace support to enos-flight-control commands (#21)
Previously, if we were downloading or unzipping a file on top of an
existing one, we would open or create it.

This works fine most of the time, however, we might instead want to fail
by default if the destination files already exists. That is now the
default behavior. A new `--replace` flag has been added to both the
download and unzip command. When given, we will now replace the
underlying file if necessary.

The replace implementation has changed from modifying the file in place to
removing and replacing it. We do this to handle cases where the files
have been `mlock()`'ed and cannot be modified.

Signed-off-by: Ryan Cragun <me@ryan.ec>