[QTI-88] Add enos_vault_start and enos_vault_init resources (#40)
* [QTI-88] Add enos_vault_start and enos_vault_init resources

Oh boy, this was quite a change but a necessary one. The purpose of this
was to build a resource that can properly start a Vault service
regardless of its version, edition, or license.

The `enos_vault_start` resource handles:
  * Creating the vault system user for running the vault service
  * Creating the configuration directory, configuration file, and
    license file for the vault service
  * Creating the systemd unit for the vault service
  * Starting the service and waiting for it to be ready for further
    actions

The `enos_vault_init` resource handles initializing the Vault cluster.

Known issues:
  * The `enos_vault_start` configuration currently requires a seal
    configuration and UI configuration even though the should be
    optional. This is due to a bug with terraform-plugin-go and
    Terraform not handling optional attributes in Objects correctly.
  * The `vault_start` configuration schema does not currently support
    everything. We'll need to expand it to support all features.

In order to provide the new functions, many general remoteflight
libraries were created to handle various things:
  * Users
  * Files
  * Systemd Units
  * Systemctl commands
  * Vault HCL files
  * Vault service management
  * Vault status management

As the libraries have been generalized we can start to refactor them
into existing resources and new resources.

The new resources also required adding support for new Terraform types
in the plugin framework. We now have an easier to use abstraction for
handling numbers, dynamic psuedo types, bools, and lists of strings.
Eventually we can also migrate to using a tfString type, a useful
intermediary until we can migrate to the terraform-plugin-framework.

Co-Authored-by: Josh Brand <jbrand@hashicorp.com>
Signed-off-by: Ryan Cragun <me@ryan.ec>