[QTI-116] migrate string state to tfString type (#42)
Support for marshaling and unmarshaling Terraform state grew organically 
from the first datasource proof-of-concept to the many stable resources
we have now. As strings are by far the most common Terraform schema
type in our resources, their implementation was first and as such is the
most naive because it was bolted onto many times as new information
was learned. Type's that followed (`tfNum`, `tfStringMap`, `tfStringSlice`,
`tfBool`, etc_) were all better because each instance carried with it the
required metadata to marshal back into the Terraform wire format without
any additional tracking elsewhere.

As migrating from every string to a `tfString `type affects every
resource, this change is quite large, but delaying this conversion any
longer will only make this a more difficult change down the road.

Now, with all state types represented as tf* types, we a common interface
that will allow us to make additional refactoring passes to automate much 
of the state transforms that are manual now.

It should also make it easier to migrate to the Terraform plugin
framework at a later date.

* Create a `tfString` type that handles marshaling and unmarshaling to a
  tftypes.String.
* Update all string collection types to handle the new `tfString` type.
* Rewrite the embedded transport merging to not require `json`.
* Create a generic `remoteflight.RunScript()` function to use wherever
  script running is required.
* Update the resource and datasource templates to use the new `tfString`
  type and fields.

Signed-off-by: Ryan Cragun <me@ryan.ec>