[QTI-308] Support unknown and dynamic values in Vault config blocks (#126)
* [QTI-308] Support unknown and dynamic values in Vault config blocks

Oh boy, this was quite the ordeal. The driving force here is the ability
to dynamically determine which unseal method to use in Vault module. In
adding support for this feature we discovered that the Vault config
blocks did not support either unknown values or inputs that didn't share
the exact same schema. To solve this we had to do several hacks to
determine Terraforms implied type of the dynamic vault config blocks. As
part of this we also upgraded to the latest version of
terraform-plugin-go and the terraform-plugin-log.

* Update terraform-plugin-log
* Update terraform-plugin-go
* Allow planning of Vault config blocks that might have unknown values

Signed-off-by: Ryan Cragun <me@ryan.ec>
Co-authored-by: Rebecca Willett <rwillett@hashicorp.com>

* please the linter

Signed-off-by: Ryan Cragun <me@ryan.ec>

* checkout before we setup go

Signed-off-by: Ryan Cragun <me@ryan.ec>

* temporarily shorten timeout

Signed-off-by: Ryan Cragun <me@ryan.ec>

* update the timeout

Signed-off-by: Ryan Cragun <me@ryan.ec>

* real bundle install needs this I guess

Signed-off-by: Ryan Cragun <me@ryan.ec>

Signed-off-by: Ryan Cragun <me@ryan.ec>
Co-authored-by: Rebecca Willett <rwillett@hashicorp.com>