[QT-563] Restructure scenario testing (#183)
* [QT-563] Restructure scenario testing

To increase confidence that we won't accidentally ship a broken provider
upstream we need to more fully test the various enos resources and
target architectures. This change collapses the ci-tests, enoscli-tests,
and and enos scenarios into a single enos directory with various
scenarios. We now have the ability to test kind, vault, vault with k8s,
and the failure handlers. Most of the scenarios have also been updated
with provider installation methods to tweak what provider installation
scource should be used. The `dev` source will use whatever is in the
local `dist` dir. This is useful for provider development, and it's also
what we use on pull request tests. The `enos` option will install the
latest enos provider from the registry. `enosdev` will install from the
`enosdev` registry.

* make: remove unused targets
* bump version
* enoscli: migrate tests into enos scenarios
* enos-ci: migrate module into enos modules
* enos: allow configuring the provider source for most scenarios
* enosdev: update vault module to use latest from
  terraform-enos-aws-vault
* enosdev: update consul module to use latest from
  terraform-enos-aws-consul
* actions: rename workflows
* actions: normalize workflow syntax
* actions: add kind scenario to all testing phases
* actions: add vault-k8s scenario to all testing phases
* actions: add vault scenario to all testing phases
* actions: add failure_handlers scenario to all testing phases
* actions: use a matrix for multiple scenarios
* actions: use different variants for the vault scenario at different
  phases.
* actions: use completed-succesfully job for branch restrictions

Signed-off-by: Ryan Cragun <me@ryan.ec>