[QT-527][QT-509] Set the vault log level in config (#187)
Add `config.log_level` to `enos_vault_start` resource to allow
configuring the log level in the `vault.hcl`. This allows operators to
`SIGHUP` the vault service to change the log level without restarting
the service. We also copy over the latest changes from
`terraform-enos-aws-vault` and `terraform-enos-aws-consul` to allow us
to configure a license for Consul Enterprise.

* enos_vault_start: add `log_level` in the `config` schema.
* version: bump a patch.
* lint: fix `mirror` error.
* enos: vendor `terraform-enos-aws-vault` module with changes for the
  consul license.
* enos: vendor `terraform-enos-aws-consul` module with changes for the
  consul license.
* enos: add support for consul license to vault module.
* actions: make vault ssh a callable workflow and call it with
  variables.
* actions: add consul edition and version to vault scenario.
* actions: normalize action step names.

Signed-off-by: Ryan Cragun <me@ryan.ec>