[QT-612] Add support for HA seals in enos_vault_start (#196)
* [QT-612] Add support for HA seals in enos_vault_start

Add support for configuring more than one prioritized seal. We do this
by making `enos_vault_start.config.seal` an optional attribute and
introduce a noew `enos_vault_start.config.seals` optional attribute. The
former is unchanged and will allow backwards compatibility. The latter
can be used to configure up to three prioritized seals. The
configuration is an object or map with three named keys that point to a
seal block configuration, i.e.

```hcl
enos_vault_start "leader" {
  config = {
    //... non-seal attributes
    seals = {
      primary = {
        type = "awskms"
        attributes = {
          kms_key_id = "..."
        }
      }
      secondary = {
        // ..
      }
      tertiary = {
        // ...
      }
    }
  }
}
```

All three seals are not required, in fact, you can use a single
`primary` to configure a single seal.

Something to keep in mind: enabling an HA seal on a Vault cluster is
complicated. You have to enable one at a time and ensure that the
cluster is fully offline and has completed its seal rewrap before
updating the configuration and restarting the cluster. The `ha_seal`
scenario in Vault is a good example of an end-to-end setup.

Along with these major changes, we also include:
* Bumping the VERSION for a new release
* Updating the Go modules
* Fixing an issue where we'd fail getting replication information for a
  vault cluster with unknown replication status.
* Gracefully handle tftypes panics and bubble up the panic from the
  provider to Terraform with diagnostics.

Signed-off-by: Ryan Cragun <me@ryan.ec>

* enable audit devices after a leader has been elected

Signed-off-by: Ryan Cragun <me@ryan.ec>

* add better error messages and validation for seals block

Signed-off-by: Ryan Cragun <me@ryan.ec>

* do more validation for the seals configuration schema

Signed-off-by: Ryan Cragun <me@ryan.ec>

---------

Signed-off-by: Ryan Cragun <me@ryan.ec>