enos_vault_start: support configuring the retry_join block in storage (#18)
* enos_vault_start: support configuring the retry_join block in storage

In order to support forcing `retry_join` to resolve using IPV6 we need
to include the address type in the `auto_join` stanza. Historically we
have handled the `retry_join` stanza in the `enos_vault_start` resource
and did not have a way to define the `retry_join` block via resource
attributes.

This does just that by supporting `retry_join` in the storage schema by
using a dynamic pseudo type for `storage` in the config.

I've tested the resource with no `retry_join` defined and added a new
variant to our test scenarios to test with it set or null. All work just
fine.

As part of this I also completely refactored how we build and test varous
provider artifacts in CI. Now we build the artifacts in parallel to speed up
the build workflow and to make it easier to only download the artifacts necessary when testing.

* enos_vault_start: add retry join variant to test scenarios
* golangci-lint: remove deprecated and disabled linters and fix new ones
* go: update deps
* version: bump version
* docs and headers
* go mod tidy
* remove implied mirror
* set exec permission after download
* actions: build artifacts in parallel
* fixup small issues with rewrite
* update actions workflows
* bump the remote host instance size and wait for cloud init to finish before running tests

Signed-off-by: Ryan Cragun <me@ryan.ec>