VAULT-30189: enos_remote_exec: Use non-static ID as prefix for scripts (#21)
* VAULT-30189: enos_remote_exec: Use non-static ID as prefix for scripts

When `enos_remote_exec` executes a script on a remote machine it copies
the contents to a local place on the target disk before executing it.
Because scripts can be re-used by different resources but provided
different env vars, we used to consider the env vars as part of the SHA
but it looks like it was it was erroneously removed[0] at some point.

What we have now is a case where multiple resources can use the same
script and therefore the same SHA in parallel. That leads to a race where
multiple resources are writing, executing, and cleaning up scripts with
the same SHA.

Instead, of reverting to the prior behavior, we instead now generate a
non-static ID for each `enos_remote_exec` resource at apply time. When
we copy over script contents we prepend this ID, along with the content
SHA, to prevent any script mod races.

[0] https://github.com/hashicorp-forge/terraform-provider-enos/commit/44fe79360ef5180d516eedc1516e4ba4e5e67742

* linter: fix lastest linter issues and pin workflows

Signed-off-by: Ryan Cragun <me@ryan.ec>