VAULT-31184: Support identity token auth in `data_artifactory_item` (#22)
The primary motivation for this PR is to add support for identity token auth to the `artifactory_item` datasource. This will allow us to migrate from API auth to identity token auth. We maintain backwards compatibility with API auth and do not change any schema, except for `username` no longer being required. If no username is provided we assume token that is required is for identity token auth, not API auth.

> [!NOTE] 
> Since this is a repo in the forge and public we cannot utilize the doormat action to test token auth in CI. As such, I've tested both the artifactory package and the datasource ACC tests locally using both identity token and API auth.

We also do a bit of housekeeping and update our Go modules and pin to the latest Github actions. 

> [!NOTE] 
> We are currently behind on `terraform-plugin-go` as the server interface now requires implementing provider functions, which we have not done yet.

* actions: pin to latest actions
* go: update deps
* bump version
* rebuild docs
* fix linter

Signed-off-by: Ryan Cragun <me@ryan.ec>