flightcontrol: build flight control for linux/s390x
Add a new build target for `flight-control` on `linux/s390x`. This is
required to use enos on such machines.

s390x is not a supported target for UPX so we cannot compress it in the
same way that we do for `linux/arm64` and `linux/amd64`, which means
that we'll have to copy ~10 MB to each target machine rather than ~3 for
the other platforms.

This will slightly increase the overall provider binary and resulting
package. Luckily the binary is zipped before upload which ought to
result in similar space savings had we been able to compress it.

* Add `linux/s390x` build target for `flight-control`
* Bump VERSION to `0.6.0`

Signed-off-by: Ryan Cragun <me@ryan.ec>