VAULT-41294: docker: handle container images that do not embed docker repository metadata (#41)
* [VAULT-41294] docker: handle container images that do not embed docker v1.1 metadata

Our CI runners recently switched to Docker 29 and that brought out all
kinds of issues we were not aware of. The list of breaking changes[0] is
long, but buried in the changes somewhere is that buildx no longer includes
the v1 repository information in the images. We had previously relied on
this our load image resource to verify that the image that we loaded
actually exists with the correct tags and ID.

Instead of relying on that, we now support fetching the same information
from the OCI metadata. It's quite a bit more complicated than the flat
information we used before but it ought to make our image loader more
robust as it can now support images created outside of Docker/moby.

By moving to the OCI metadata we also have to have to handle changes in
containerd's archive import (which is used by Kind) when using images
that don't have older Docker metadata. The archive importer will always
normalize repository references to include a domain (docker.io) when the
images are not created with it. When searching for images with a
reference that does not include a domain we'll attempt to find matching
images with the default domain if necessary.

As part of this I also ran the latest gofumpt and golancilint.

[0]: https://docs.docker.com/engine/release-notes/29/

Signed-off-by: Ryan Cragun <me@ryan.ec>